Added variables: REGENFOOD, BONUSHITSMAX, BONUSMANAMAX, BONUSSTAMMAX variables to prepare for the overhaul of the status bar packet.
Moved: All shared variables of CItem and CChar were combined into CObjBase. I would like to place a warning some MIGHT need to get moved back depending the status bar packet.

NOTE: the variables for the REGEN* should possibly be split again. As the Item should Add to the Char. Items code SHOULD be soft coded to affect the change in the Char variable for the overrides in REGEN*.