Added initial support for MultiCollection.uop multi format (still incomplete) MultiCollection.uop support is required to use new TOL multis, which includes many new keeps/castles
This commit is only to make Sphere detect when MultiCollection.uop is found on UO folder to make it override old multi.mul + multi.idx. But the code is commented out, still need to find a way to read MultiCollection.uop data properly

PS: I don't know how to deal with UO file reading so maybe someday this thing will work, or maybe not. If someone want try it, these OrionUO commits should help with something:
https://github.com/Hotride/OrionUO/commit/70eab08d36080cce310b1b072397e84d0a71ad5a
https://github.com/Hotride/OrionUO/commit/7931745586e70e2e31a78f83fcbc418fd6e16c65